Support for Rv64

+++ src/cmdline.rs
@@ -15,8 +15,8 @@ pub struct Arguments {
     pub filename: String,
     pub exe_option: ExeOption,
     pub pkpath: Option<String>,
-    pub init_pc: Option<u32>,
-    pub break_point: Option<u32>,
+    pub init_pc: Option<u64>,
+    pub break_point: Option<u64>,
     pub result_reg: Option<usize>,
     pub main_args: Option<Vec<String>>,
 }
@@ -70,12 +70,12 @@ impl Arguments {
         };
 
         let init_pc = app.value_of("pc").map(|x| {
-            u32::from_str_radix(x.trim_start_matches("0x"), 16)
+            u64::from_str_radix(x.trim_start_matches("0x"), 16)
                 .expect("invalid pc\nplease set value as hex (e.g. --pc=0x80000000)")
         });
 
         let break_point = app.value_of("break_point").map(|x| {
-            u32::from_str_radix(x.trim_start_matches("0x"), 16)
+            u64::from_str_radix(x.trim_start_matches("0x"), 16)
                 .expect("invalid break point\nplease set value as hex (e.g. --pc=0x80000000)")
         });
